Change hardcoded version of nginx-accesskey module to master branch.
nginx-accesskey hardcoded version is 2.0.3 but the latest one is 2.0.4.
Now nginx-accesskey is downloaded from master branch.

Signed-off-by: alexgoussev <algous@gmail.com>

typo

Change hardcoded version of nginx-accesskey module to master one.

nginx-accesskey hardcoded version is 2.0.3 but the latest one is 2.0.4.
It is better to download master branch. So that it is no need to track version changes.
Now nginx-accesskey is downloaded from master branch.