update tools/cminfo.sh for centos 6 in 123.09beta01
Updated compatibility fix for centos 6 systems which do not have datamash YUM package so can not display more extended sar cpu utilisation and memory usage stats for min, max and 95% percentile values in additional to averages. So this update makes sure to check for centos 6 and fall back to more average displayed metrics.